fix: move some function from post to profile
